requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.5.0
appdirs==1.4.4
asyncpg==0.29.0
certifi==2024.8.30
click==8.1.7
dictdiffer==0.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
iso8601==1.1.0
jedi==0.19.1
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
parso==0.8.4
prompt_toolkit==3.0.47
ptpython==3.0.29
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
rich==13.8.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.38.5
tomlkit==0.13.2
tortoise-cli==0.1.2
tortoise-orm==0.21.6
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
